requirements.txt
streamlit
fastapi
uvicorn
pandas
numpy
qdrant-client
sentence-transformers
transformers
torch
scikit-learn
python-dotenv
pytest
